fix: codigo organizado em componentes
